Correctly handle POLLRDHUP registration in all cases
Motivation:

When accepting a Channel we did register it for POLLRDHUP, but unfortunally we used the IOUringSubmissionQueue that is tied to the IOUringEventLoop that is used for the ServerChannel. This is not correct as the EventLoop used for the accepted Channel may be different.

Modification:

Move logic into doRegister() and so register for POLLRDHUP on the right IOURingSubmissionQueue

Result:

Correct POLLRDHUP handling