Add docker files to compile io_uring transport (#7)
Motivation:

Add docker files that can be used to setup the right env to compile the transport. This can also be used to release as it will ensure we link against the minimum GLIBC version and so make it usable everywhere as long as a recent enough kernel is present.

Modifications:

Add docker files

Result:

Be able to compile easily and release