nextWakupNanos lazySet.
The memory ordering effects of the nextWakupNanos are irrelevant. We
just need to make sure that the compiler doesn't optimize out the load
or the store. A setOpaque should be sufficient. But since the code needs
to be Java 8 compatible, setRelease is the cheapest alternative.

This will restrict compiler optimizations but doesn't lead to a memory
fence on the X86 since every store is a release store (the semantics
provided by a setLazy).

In a subsequent PR I'll deal with the getAndSet in the finally clause.