Add build files for MacOS/M1
Motivation:
We're increasingly developing on MacOS/M1 machines, so it's useful to have a standard set-up for building and testing locally.

Modification:
Add another set of docker files for building native ARM (not cross-compiling) binaries.
Add a Makefile to make it easy to start these builds or drop into the docker guest shell.

Result:
Easier development flow when working on a MacOS/M1 machine.