Rewrite the transport for Netty 5
This is a first draft of rewriting the transport to work with Netty 5.
There are many changes; IoHandler, AbstractChannel, Buffer, ... all very different in Netty 5.
The io_uring JNI integration code remain largely the same.

Another change is being attempted in this rewrite: to use blocking file descriptors and rely on FEAT_FAST_POLL,
where blocking IO operations are handled efficiently with a poll-arming mechanism inside the kernel.
This means we don't have to issue separate poll-in/-out and read/write operations; we can instead just
submit our IO operations and wait for them to complete.

The state of this is that basic socket IO works. Datagrams and domain sockets are broken or missing,
TCP fast-open is missing, SSL handshake times out, many tests are failing or timing out.
But many tests are also passing, so progress is being made.