Minor allocation optimization
At the moment, for every multiple write operation an iov array buffer is allocated. Which cause some churn when there are high volume of writes. Considering at max there can be `Limits.IOV_MAX` values, creating the ByteBuf once and re-using it based on the elements in the schedule operation.