Add devcontainers for Linux
Motivation:
Linux is the most important deployment OS, but a significant amount of
development takes place on Windows and MacOS. Devcontainers allow
developers to quickly spin up a Docker container, to run their IDE
inside a native Linux environment. This is helpful for debugging and
testing.
Intellij IDEA Ultimate and Visual Studio Code both have support for
devcontainers.

Modifications:
Add devcontainer spec file for ubuntu

Result:
It's now possible to run an IDE in a Linux devcontainer environment on
Windows and MacOS.