Log submission failures via tracing level as these are expected (#263)
Motivation:

Sometimes submission of ops will fail if these are executed inline and
return some error. This is expected and so we shouldn't log a warn
message. That said it might still be useful to log these cases with
trace level.

Modifications:

Log submission failures via tracing level

Result:

Use correct logging level that will not confuse people. This is a port
of https://github.com/netty/netty/pull/14671